requirements.txt
pyaudio
SpeechRecognition
gTTS
python-dotenv
streamlit
pipwin

-e .